Moves components related to drafts into modules/drafts
