+++ src/frontend/web_application/src/modules/drafts/components/DiscussionDraft/index.spec.jsx
