+++ src/frontend/web_application/src/modules/drafts/components/DiscussionTextarea/presenter.jsx
