+++ src/frontend/web_application/src/modules/drafts/components/DiscussionTextarea/style.scss
@@ -1,5 +1,5 @@
-@import '../../styles/common';
-@import '../../styles/object/o-form-element';
+@import '../../../../styles/common';
+@import '../../../../styles/object/o-form-element';
 
 .m-discussion-textarea {
   &__body {
