+++ src/frontend/web_application/src/modules/drafts/components/DraftMessageActionsContainer/presenter.jsx
@@ -2,9 +2,8 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import { Trans } from 'lingui-react';
-import Button from '../Button';
-import Modal from '../Modal';
-import { ManageEntityTags } from '../../modules/tags';
+import { Button, Modal } from '../../../../components';
+import { ManageEntityTags } from '../../../../modules/tags';
 
 import './style.scss';
