+++ src/frontend/web_application/src/modules/drafts/components/DraftMessageActionsContainer/style.scss
@@ -1,4 +1,4 @@
-@import '../../styles/vendor/bootstrap_foundation-sites';
+@import '../../../../styles/vendor/bootstrap_foundation-sites';
 
 .m-message-actions-container {
