+++ src/frontend/web_application/src/modules/drafts/components/NewDraftForm/index.js
