+++ src/frontend/web_application/src/modules/drafts/components/NewDraftForm/presenter.jsx
@@ -2,12 +2,8 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { v1 as uuidV1 } from 'uuid';
 import { Trans } from 'lingui-react';
-import Button from '../Button';
-import Icon from '../Icon';
-import Spinner from '../Spinner';
-import { Dropdown, withDropdownControl } from '../Dropdown';
-import TextFieldGroup from '../TextFieldGroup';
-import ContactAvatarLetter from '../ContactAvatarLetter';
+import { Button, Icon, Spinner, Dropdown, withDropdownControl, TextFieldGroup } from '../../../../components/';
+import ContactAvatarLetter from '../../../../components/ContactAvatarLetter';
 import DiscussionDraft, { TopRow, BodyRow, BottomRow } from '../DiscussionDraft';
 import DiscussionTextarea from '../DiscussionTextarea';
 import RecipientList from '../RecipientList';
