+++ src/frontend/web_application/src/modules/drafts/components/Recipient/index.js
