+++ src/frontend/web_application/src/modules/drafts/components/RecipientList/presenter.spec.jsx
