+++ src/frontend/web_application/src/modules/drafts/components/ReplyForm/presenter.jsx
@@ -2,9 +2,8 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { v1 as uuidV1 } from 'uuid';
 import { Trans } from 'lingui-react';
-import { Button, Spinner, Icon, Dropdown, withDropdownControl } from '../';
-import Link from '../Link';
-import ContactAvatarLetter from '../ContactAvatarLetter';
+import { Link, Button, Spinner, Icon, Dropdown, withDropdownControl } from '../../../../components';
+import ContactAvatarLetter from '../../../../components/ContactAvatarLetter';
 import DiscussionDraft, { TopRow, BodyRow, BottomRow } from '../DiscussionDraft';
 import DiscussionTextarea from '../DiscussionTextarea';
 import './style.scss';
