+++ src/frontend/web_application/src/modules/drafts/components/ReplyForm/style.scss
@@ -1,5 +1,5 @@
-@import '../../styles/vendor/bootstrap_foundation-sites';
-@import '../../styles/object/o-avatar';
+@import '../../../../styles/vendor/bootstrap_foundation-sites';
+@import '../../../../styles/object/o-avatar';
 
 .m-reply {
   @include flex-grid-row($size: expand);
